requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient==2.1.1
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
